[Fix] 팩토리 메소드 생성
Piece의 변경으로 인해 깨진 테스트 코드 및 Board 클래스를 수리
