+++ src/test/java/pe/goblin/chess/board/BoardTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import pe.goblin.chess.pawn.Piece;
+import pe.goblin.chess.piece.Piece;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -26,12 +26,12 @@ public class BoardTest {
     @Test
     @DisplayName("Pawn을 체스 판에 추가할 수 있다.")
     void create() throws ExceedPawnException {
-        Piece white = new Piece(Piece.WHITE_COLOR, Piece.WHITE_REPRESENTATION);
+        Piece white = Piece.createWhitePawn();
         board.add(white);
         assertEquals(1, board.size());
         assertEquals(white, board.findPawn(0));
 
-        Piece black = new Piece(Piece.BLACK_COLOR, Piece.WHITE_REPRESENTATION);
+        Piece black = Piece.createBlackPawn();
         board.add(black);
         assertEquals(2, board.size());
         assertEquals(black, board.findPawn(1));
@@ -44,18 +44,18 @@ public class BoardTest {
         @DisplayName("white pawn을 추가하는 경우")
         void in_case_of_white_pawn() throws ExceedPawnException {
             for (int i = Board.MIN_ROWS; i < Board.MAX_ROWS; i++) {
-                board.add(new Piece(Piece.WHITE_COLOR, Piece.WHITE_REPRESENTATION));
+                board.add(Piece.createWhitePawn());
             }
-            assertThrows(Exception.class, () -> board.add(new Piece(Piece.WHITE_COLOR, Piece.WHITE_REPRESENTATION)));
+            assertThrows(Exception.class, () -> board.add(Piece.createWhitePawn()));
         }
 
         @Test
         @DisplayName("black pawn을 추가하는 경우")
         void in_case_of_black_pawn() throws ExceedPawnException {
             for (int i = Board.MIN_ROWS; i < Board.MAX_ROWS; i++) {
-                board.add(new Piece(Piece.BLACK_COLOR, Piece.BLACK_REPRESENTATION));
+                board.add(Piece.createBlackPawn());
             }
-            assertThrows(Exception.class, () -> board.add(new Piece(Piece.BLACK_COLOR, Piece.BLACK_REPRESENTATION)));
+            assertThrows(Exception.class, () -> board.add(Piece.createBlackPawn()));
         }
     }
